credential

+++ app/GoogleSheet.php
@@ -16,12 +16,15 @@ class GoogleSheet {
         // //$client->setAuthConfig($credential);
         // return new \Google_Service_Sheets($client);
         
+        $credentials_file = storage_path('app/json/lev-college-43ca656233f1.json');
+        if(! $file_exists($credentials_file)) {
+            file_put_contents('credentials.json', getenv('GOOGLE_CREDENTIALS'));
+            $credentials_file = 'credentials.json';
+        }
         
-        file_put_contents('credentials.json', getenv('GOOGLE_CREDENTIALS'));
-        //$credentials_path = env('GOOGLE_CREDENTIALS');
         $client = new \Google_Client();
         $client->setScopes([[\Google_Service_Sheets::SPREADSHEETS]]);
-        $client->setAuthConfig('credentials.json');
+        $client->setAuthConfig($credentials_file);
         return new \Google_Service_Sheets($client);
